leptos_state::compat

Struct ReadSignal

pub struct ReadSignal<T, S = SyncStorage> { /* private fields */ }

An arena-allocated getter for a reactive signal.

A signal is a piece of data that may change over time, and notifies other code when it has changed.

This is an arena-allocated signal, which is Copy and is disposed when its reactive Owner cleans up. For a reference-counted signal that lives as long as a reference to it is alive, see [ArcReadSignal].

Core Trait Implementations

.get() clones the current value of the signal. If you call it within an effect, it will cause that effect to subscribe to the signal, and to re-run whenever the value of the signal changes.

.get_untracked() clones the value of the signal without reactively tracking it.

.read() returns a guard that allows accessing the value of the signal by reference. If you call it within an effect, it will cause that effect to subscribe to the signal, and to re-run whenever the value of the signal changes.

.read_untracked() gives access to the current value of the signal without reactively tracking it.

.with() allows you to reactively access the signal’s value without cloning by applying a callback function.

.with_untracked() allows you to access the signal’s value by applying a callback function without reactively tracking it.

.to_stream() converts the signal to an async stream of values.

::from_stream() converts an async stream of values into a signal containing the latest value.

Examples

let (count, set_count) = signal(0);

// calling .get() clones and returns the value
assert_eq!(count.get(), 0);
// calling .read() accesses the value by reference
assert_eq!(count.read(), 0);